Fix a minor bug in the way the values were stored

+++ reaction-diffusion/include/utils.hpp
@@ -6,8 +6,8 @@
 
 int random_int(int min, int max);
 
-float laplaceA(std::vector<std::vector<std::pair<int, int>>> *grid, int i, int j);
+float laplaceA(std::vector<std::vector<std::pair<float, float>>> *grid, int i, int j);
 
-float laplaceB(std::vector<std::vector<std::pair<int, int>>> *grid, int i, int j);
+float laplaceB(std::vector<std::vector<std::pair<float, float>>> *grid, int i, int j);
 
 #endif
